add new tool: etvdb_cli
this is as of now the etvdb simple/synchronous api reference
implementation.
as the api works fine for this usecase, it will probably stay,
an asynchronous api may follow, but will get new functions or a new
namespace.

for now this only outputs the requested data in a CSV-like format,
in future there will be a file mode to rename files and more.